# swagger.json
swagger: '2.0'
info:
  title: Healthcare System API
  description: API for managing healthcare system operations
  termsOfService: https://www.example.com/policies/terms/
  contact:
    email: contact@example.com
  license:
    name: BSD License
  version: v1
host: 127.0.0.1:8000
schemes:
- http
basePath: /api
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Basic:
    type: basic
security:
- Basic: []
paths:
...
  /auth/login/:
    post:
      operationId: auth_login_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/TokenObtainPair'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/TokenObtainPair'
      tags:
      - auth
    parameters: []
...
  /patients/:
    get:
      operationId: patients_list
      description: ''
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            type: array
            items:
              $ref: '#/definitions/Patient'
      tags:
      - patients
    post:
      operationId: patients_create
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/Patient'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/Patient'
      tags:
      - patients
    parameters: []
...
definitions:
  DoctorList:
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      preferred_name:
        title: Preferred name
        type: string
        maxLength: 255
...
  TokenObtainPair:
    required:
    - username
    - password
    type: object
    properties:
      username:
        title: Username
        type: string
        minLength: 1
      password:
        title: Password
        type: string
        minLength: 1
...
  Patient:
    required:
    - first_name
    - last_name
    - national_id
    - birth_date
    - gender
    - phone
    - modality
    - status
    - hospital
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      first_name:
        title: First name
        type: string
        maxLength: 100
        minLength: 1
      last_name:
        title: Last name
        type: string
        maxLength: 100
        minLength: 1
      national_id:
        title: National id
        type: string
        maxLength: 20
        minLength: 1
      birth_date:
        title: Birth date
        type: string
        format: date
      age:
        title: Age
        type: integer
        maximum: 2147483647
        minimum: -2147483648
        x-nullable: true
      gender:
        title: Gender
        type: string
        enum:
        - male
        - female
      phone:
        title: Phone
        type: string
        maxLength: 20
        minLength: 1
      modality:
        title: Modality
        type: string
        maxLength: 50
        minLength: 1
      status:
        title: Status
        type: string
        maxLength: 20
        minLength: 1
      hospital:
        title: Hospital
        type: integer
      reference_doctor:
        title: Reference doctor
        type: integer
        x-nullable: true
      related_doctor:
        title: Related doctor
        type: integer
        x-nullable: true
      reference_doctor_data:
        $ref: '#/definitions/DoctorList'
      registration_date:
        title: Registration date
        type: string
        format: date-time
      visit_date:
        title: تاریخ ویزیت
        type: string
        format: date
        x-nullable: true
      doctor:
        $ref: '#/definitions/DoctorList'
...